requirements.txt
# This file is intentionally left empty as no external libraries are required.
# Python version requirement: >=3.7

Flask==2.1.0
Jinja2==3.0.3
Markdown==3.3.6
python-frontmatter==1.0.0